chore: Rearrange the telemetry package and ignore some logging
